Add unscaled decision trees
NOTE: Weird happenings re: the k = 4 pruned tree. Seems like in the
previous commit (#a0250ee) k = 4 was actually an unpruned tree.
Rerunning the script with SAVE.DTREES = TRUE and pdflatex fixes this.
I'm going to push a rerun of pdflatex soon, which then can be compared
to the previous version - not sure what happened.